feat(helpers): Add gqlFetchAll to AppToolkit (#792)
