+++ src/app-toolkit/helpers/the-graph/the-graph.helper.ts
@@ -16,6 +16,16 @@ type RequestGraphParams = {
   headers?: Record<string, string>;
 };
 
+interface gqlFetchAllParams<T> {
+  query: string;
+  endpoint: string;
+  variables?: Variables;
+  dataToSearch: string;
+  offset?: number;
+  first?: number;
+  prevResults?: T;
+}
+
 @Injectable()
 export class TheGraphHelper {
   async request<T = any>({ endpoint, query, variables, headers }: RequestParams) {
@@ -23,11 +33,68 @@ export class TheGraphHelper {
     return client.request<T>(query, variables);
   }
 
-  async requestGraph<T>({ endpoint, query, variables = {} }: RequestGraphParams) {
+  async requestGraph<T = any>({ endpoint, query, variables = {} }: RequestGraphParams) {
     const presentQuery = typeof query === 'string' ? query : query.present;
     const pastQuery = typeof query === 'string' ? null : query.past;
     const finalQuery = typeof variables.blockTag === 'number' && pastQuery ? pastQuery : presentQuery;
 
     return this.request<T>({ endpoint, query: finalQuery, variables });
+  }
+
+  async gqlFetchAll<T = any>({
+    query,
+    endpoint,
+    variables = {},
+    dataToSearch,
+    offset = 0,
+    first = 1000,
+    prevResults,
+  }: gqlFetchAllParams<T>): Promise<T> {
+    const results = await this.requestGraph<T>({
+      endpoint,
+      query,
+      variables: {
+        ...variables,
+        first: first,
+        skip: offset,
+      },
+    });
+
+    if (results[dataToSearch].length === first) {
+      let newPrevResults = results;
+      if (prevResults) {
+        newPrevResults = {
+          ...prevResults,
+          ...results,
+          [dataToSearch]: [
+            ...prevResults[dataToSearch],
+            ...results[dataToSearch],
+          ],
+        };
+      }
+
+      return this.gqlFetchAll({
+        query,
+        endpoint,
+        variables,
+        dataToSearch,
+        first: first,
+        offset: offset + first,
+        prevResults: newPrevResults,
+      });
+    }
+
+    if (prevResults) {
+      return {
+        ...prevResults,
+        ...results,
+        [dataToSearch]: [
+          ...prevResults[dataToSearch],
+          ...results[dataToSearch],
+        ],
+      };
+    } else {
+      return results;
+    }
   }
 }
